Categories CRUD

+++ Modules/Product/database/migrations/2023_12_28_141106_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name')->unique();
             $table->string('slug');
             $table->string('description')->nullable();
-            $table->decimal('price',5,2);
+            $table->float('price',5,2);
             $table->unsignedBigInteger('categories_id');
             $table->timestamps();
             // $table->foreign('categories_id')->references('id')->on('category');
